Quick Push before sleep
