+++ help/Color-Configurations/guide.md
@@ -1,0 +1,5 @@
+# Color Configurations!
+___
+I'm assuming that you might have stumbled over here by pure curiosity or not, anyways lets get to it!
+
+First off open
