+++ help/installing/guide.md
@@ -1,16 +1,16 @@
 # Installing
-## Please be careful to be in the Downloads directory when installing otherwise it wont work!
+## Please be careful to install gcc if on Ubuntu before trying to install
 ___
 
 Installing Note is a really easy task to do!
 I made a quick little script to install the latest version of Note for anyone who wishes to use note but doesn't know how to install with source code!
 First off to start you are going to want to run these commands in your terminal
 
-    curl https://raw.githubusercontent.com/Awesome-Sauces/Note/main/installing/ArchInstaller.sh > ArchInstaller.sh
+    curl https://raw.githubusercontent.com/Awesome-Sauces/Note/main/help/installing/resources/ArchInstaller.sh > ArchInstaller.sh
 
 For Ubuntu do this
 
-    wget https://raw.githubusercontent.com/Awesome-Sauces/Note/main/installing/DebianInstaller.sh
+    wget https://raw.githubusercontent.com/Awesome-Sauces/Note/main/help/installing/resources/DebianInstaller.sh
 
 Replace the raw file link with your Linux Operating system Installer.
 For this example we will use the ArchLinux Installer,
