+++ help/installing/resources/ArchInstaller.sh
