+++ help/installing/resources/DebianInstaller.sh
@@ -1,6 +1,5 @@
 #!/bin/bash
 pkg="go"
-u="$USER"
 which $pkg > /dev/null 2>&1
 if [ $? == 0 ]
 then
@@ -10,6 +9,8 @@ wget https://raw.githubusercontent.com/Awesome-Sauces/Note/main/src/note.go
 wget https://raw.githubusercontent.com/Awesome-Sauces/Note/main/src/go.mod
 go build note.go
 sudo mv note /usr/bin
+rm note.go
+rm go.mod
 else
     read -p "Golang is not installed. Would you like to install it? y/n : " request
 if  [ $request == "y" ]
@@ -19,10 +20,10 @@ then
 (cd /tmp; tar -xzvf go1.17.linux-amd64.tar.gz)
 (cd /tmp; sudo mv go /usr/local)
 (cd /tmp; rm -rf go)
-echo 'export GOROOT=/usr/local/go' >> /home/$u/.bashrc
-echo 'export GOPATH=$HOME/go' >> /home/$u/.bashrc
-echo 'export PATH=$GOPATH/bin:$GOROOT/bin:$PATH' >> /home/$u/.bashrc
-source /home/$u/bashrc
+echo 'export GOROOT=/usr/local/go' >> ~/.bashrc
+echo 'export GOPATH=$HOME/go' >> ~/.bashrc
+echo 'export PATH=$GOPATH/bin:$GOROOT/bin:$PATH' >> ~/.bashrc
+source ~/.bashrc
 # Grabs Newest version of Note
 wget https://raw.githubusercontent.com/Awesome-Sauces/Note/main/src/note.go
 wget https://raw.githubusercontent.com/Awesome-Sauces/Note/main/src/go.mod
