+++ src/colorConfig.json
@@ -1,25 +1,25 @@
 {
-    "TextColor": [
-      {
-        "ColorRed":"false",
-        "ColorGreen": "false",
-        "ColorYellow": "false",
-        "ColorBlue": "false",
-        "ColorPurple": "true",
-        "ColorCyan" : "false",
-        "ColorWhite" : "true"
-      }
-    ],
-    "CursorColor": [
-  
-      {
-        "ColorRed":"false",
-        "ColorGreen": "false",
-        "ColorYellow": "false",
-        "ColorBlue": "false",
-        "ColorPurple": "true",
-        "ColorCyan" : "false",
-        "ColorWhite" : "false"
-      }
-    ]
-  }
+  "TextColor": [
+    {
+      "ColorRed":"false",
+      "ColorGreen": "false",
+      "ColorYellow": "false",
+      "ColorBlue": "false",
+      "ColorPurple": "false",
+      "ColorCyan" : "false",
+      "ColorWhite" : "true"
+    }
+  ],
+  "CursorColor": [
+
+    {
+      "ColorRed":"false",
+      "ColorGreen": "false",
+      "ColorYellow": "false",
+      "ColorBlue": "false",
+      "ColorPurple": "true",
+      "ColorCyan" : "false",
+      "ColorWhite" : "false"
+    }
+  ]
+}
